error msgs for builtins
